feat: moved circuit deleting cheching to vue component

+++ src/components/TabsBar/TabsBar.vue
@@ -42,17 +42,17 @@
         </button>
     </div>
     <MessageBox
-        v-model="showMessage"
+        v-model="dispMessage"
         :button-list="buttonArr"
         :is-persistent="persistentShow"
         :message-text="messageVal"
-        @hide-message="showMessage = false"
+        @hide-message="dispMessage = false"
     />
 </template>
 
 <script lang="ts" setup>
 import draggable from 'vuedraggable'
-import { truncateString } from '#/simulator/src/utils'
+import { showMessage, truncateString } from '#/simulator/src/utils'
 import { ref } from '@vue/reactivity'
 import { computed, onMounted } from '@vue/runtime-core'
 import {
@@ -65,7 +65,7 @@ import MessageBox from '#/components/MessageBox/messageBox.vue'
 
 const drag = ref(false)
 const updateCount = ref(0)
-const showMessage = ref(false)
+const dispMessage = ref(false)
 const persistentShow = ref(false)
 const messageVal = ref('')
 const buttonArr = ref([''])
@@ -75,33 +75,44 @@ function closeCircuit(e, circuitItem) {
 
     // check circuit count
     if (SimulatorStore().circuit_list.length <= 1) {
-        showMessage.value = true
+        dispMessage.value = true
         persistentShow.value = false
         messageVal.value =
             'At least 2 circuits need to be there in order to delete a circuit.'
         buttonArr.value = ['close']
         return
     }
-    showMessage.value = false
+    dispMessage.value = false
     persistentShow.value = false
     buttonArr.value = []
 
     let dependencies = getDependenciesList(circuitItem.id)
     if (dependencies) {
         dependencies = `\nThe following circuits are depending on '${circuitItem.name}': ${dependencies}\nDelete subcircuits of ${circuitItem.name} before trying to delete ${circuitItem.name}`
-        showMessage.value = true
+        dispMessage.value = true
         persistentShow.value = true
         messageVal.value = dependencies
         buttonArr.value = ['OK']
         return
     }
 
-    console.log(circuitItem)
-    var index = SimulatorStore().circuit_list.indexOf(circuitItem)
-    if (index !== -1) {
-        SimulatorStore().circuit_list.splice(index, 1)
+    const confirmation = confirm(
+        `Are you sure want to close: ${circuitItem.name}\nThis cannot be undone.`
+    )
+
+    if (confirmation) {
+        var index = SimulatorStore().circuit_list.indexOf(circuitItem)
+        if (index !== -1) {
+            SimulatorStore().circuit_list.splice(index, 1)
+        }
+        deleteCurrentCircuit(circuitItem.id)
+        showMessage('Circuit was successfully closed')
+    } else {
+        showMessage('Circuit was not closed')
     }
-    deleteCurrentCircuit(circuitItem.id)
+
+    console.log(circuitItem)
+
     updateCount.value++
 }
